removing as a colaborator and more comments

+++ models/documents.js
@@ -13,6 +13,7 @@ const docs = {
     //Gets all the documents
     getAll: async (userEmail) => {
         const db = await database.getDb('documents');
+        console.log('connected to database');
         return await db.collection.find({ 
             $or: [
                 {userEmail},
@@ -25,6 +26,7 @@ const docs = {
     getOne: async (id, userEmail) => {
         const db = await database.getDb('documents');
         const docId = new mongo.ObjectId(id);
+        console.log(`founf document ${docId}`);
         return await db.collection.findOne({ 
             _id: docId, 
             $or: [
@@ -41,9 +43,9 @@ const docs = {
             ...document,
             userEmail,
             createdAt: new Date(),
-            type: document.type || 'text'
         };
         const response = await db.collection.insertOne(newDocument);
+        console.log(`New document created with ID: ${response.insertedId}`)
         return response.insertedId;
     },
 
@@ -51,6 +53,7 @@ const docs = {
     update: async (id, updatedData, userEmail) => {
         const db = await database.getDb('documents');
         const docId = new mongo.ObjectId(id);
+        console.log(`updating document ${docId}`);
         await db.collection.findOneAndUpdate(
             { _id: docId, $or: [{userEmail}, {collaborator: userEmail}] },
             { $set: updatedData },
@@ -58,11 +61,21 @@ const docs = {
         );
     },
 
-    //Deletes a document
+    //Deletes a document from userEmail (owner or collaborator)
     delete: async (id, userEmail) => {
-        const db = await database.getDb('documents');
+        const {db, collection} = await database.getDb('documents');
         const docId = new mongo.ObjectId(id);
-        await db.collection.findOneAndDelete({ _id: docId, userEmail });
+        const document = await collection.findOne({_id: docId})
+        if (document.userEmail === userEmail) {
+            await db.collection.findOneAndDelete({ _id: docId, userEmail });
+            console.log('Document has been removed')
+        } else {
+            await collection.updateOne(
+                {_id: docId},
+                {$pull: {collaborator: userEmail}}
+            );
+            console.log('You are have been removed from the document as a collaborator!');
+        }
     },
 
     //Shares a document to another user
@@ -72,38 +85,42 @@ const docs = {
         const document = await collection.findOne({ _id: docId, userEmail});
 
         if (!document) {
-            throw new Error("document not found");
+            console.log("document not found");
         }
 
         const userCollection = db.collection('users');
         const collaborator = await userCollection.findOne({email});
 
-        //Checks if the invited user already exists
         if (collaborator) {
+            const invitationUrl = `http://localhost:3000/`;
+            const msg = {
+                to: email,
+                from: 'anteo.ssr@gmail.com',
+                subject: `${userEmail} has added you as a collaborator`,
+                text: `You've added as a collaborator on a document. Login using this link: ${invitationUrl}`,
+            }
+
+            await sgMail.send(msg);
             await collection.updateOne(
                 { _id: docId },
                 { $addToSet: {collaborator: email}}
             );
-            
-            return {message: "Document shared"};
+            console.log('Document shared');
         } else {
             const invitationUrl = `http://localhost:3000/#/register?email=${email}`;
-
             const msg = {
                 to: email,
                 from: 'anteo.ssr@gmail.com',
                 subject: `invitation from ${userEmail}`,
-                text: `You've been invited to collaborate on a document. Please register using this link: ${invitationUrl}`,
+                text: `You've been invited to collaborate on a document. Register using this link: ${invitationUrl}`,
             }
 
             await sgMail.send(msg);
-
             await collection.updateOne(
                 { _id: docId },
                 { $addToSet: { collaborator: email } }
             );
-
-            return {message: 'invitation sent'}
+            console.log('Document shared');
         }
     },
 
@@ -122,9 +139,9 @@ const docs = {
 
         if (!response.ok) {
             const errorResponse = await response.json();
-            throw new Error(`Execution failed: ${errorResponse.message}`);
+            console.log(`Execution failed: ${errorResponse.message}`);
         }
-
+        console.log('data processed');
         return JSON.parse(textResponse);
     }
 };
